requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
PyPDF2==3.0.1
youtube-transcript-api==0.6.2
pytube==15.0.0
sentence-transformers==2.2.2
chromadb==0.4.15
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
arxiv==1.4.8
ollama==0.1.7
